+++ clippy_lints/src/lib.register_lints.rs
@@ -47,7 +47,6 @@ store.register_lints(&[
     attrs::DEPRECATED_SEMVER,
     attrs::EMPTY_LINE_AFTER_OUTER_ATTR,
     attrs::INLINE_ALWAYS,
-    attrs::MISMATCHED_TARGET_OS,
     attrs::USELESS_ATTRIBUTE,
     await_holding_invalid::AWAIT_HOLDING_LOCK,
     await_holding_invalid::AWAIT_HOLDING_REFCELL_REF,
@@ -75,6 +74,7 @@ store.register_lints(&[
     casts::FN_TO_NUMERIC_CAST_WITH_TRUNCATION,
     casts::PTR_AS_PTR,
     casts::UNNECESSARY_CAST,
+    cfg_target_os::MISMATCHED_TARGET_OS,
     checked_conversions::CHECKED_CONVERSIONS,
     cognitive_complexity::COGNITIVE_COMPLEXITY,
     collapsible_if::COLLAPSIBLE_ELSE_IF,
